requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
Deprecated==1.2.13
Django==4.1.2
django-filter==22.1
django-redis==5.2.0
django-seed==0.3.1
djangorestframework==3.14.0
Faker==15.1.1
importlib-metadata==5.0.0
Markdown==3.4.1
packaging==21.3
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
redis==4.3.4
six==1.16.0
sqlparse==0.4.3
toposort==1.7
wrapt==1.14.1
zipp==3.10.0
